+++ src/eca/llm_providers/anthropic.clj
@@ -9,7 +9,7 @@
    [eca.llm-util :as llm-util]
    [eca.logger :as logger]
    [eca.oauth :as oauth]
-   [eca.shared :as shared :refer [assoc-some deep-merge multi-str]]
+   [eca.shared :as shared :refer [assoc-some deep-merge join-api-url multi-str]]
    [hato.client :as http]
    [ring.util.codec :as ring.util]))
 
@@ -69,7 +69,7 @@
                       :cache_control {:type "ephemeral"}})))
 
 (defn ^:private base-request! [{:keys [rid body api-url api-key auth-type url-relative-path content-block* on-error on-stream http-client extra-headers]}]
-  (let [url (str api-url (or url-relative-path messages-path))
+  (let [url (join-api-url api-url (or url-relative-path messages-path))
         reason-id (str (random-uuid))
         oauth? (= :auth/oauth auth-type)
         headers (client/merge-llm-headers
@@ -403,9 +403,9 @@
   (swap! db* assoc-in [:auth provider] {:step :login/waiting-login-method})
   (send-msg! (multi-str "Now, inform the login method:"
                         ""
-                        "- max: Claude Pro/Max"
-                        "- console: Automatically create API Key and use it"
-                        "- manual: Manually enter API Key")))
+                        "- max: Claude Pro/Max (for claude.ai accounts, subscription)"
+                        "- console: Automatically create API Key and use it (non subscription)"
+                        "- manual: Manually enter API Key (non subscriptions)")))
 
 (defmethod f.login/login-step ["anthropic" :login/waiting-login-method] [{:keys [db* input provider send-msg!]}]
   (case input